Temporary schedule set up - final order in flux
Order to teach R and SQL in, is optional. Many find that teaching spreadsheets > open refine > sql > r is a preferred logical order. Others have incorporated showing how SQL fits in the Data analysis and visualization in R section at the end of the second day. In this latter case, intro to R is the second part of day 1. Data analysis and visualization in R is at the beginning of day 2 and SQL is after that. Stay tuned. @mjcollin will likely teach the R / R analysis and visualization section and will have a preference for where to put SQL.